Add PrimProc support for selective block loading for 16-byte columns.

+++ dbcon/joblist/batchprimitiveprocessor-jl.h
@@ -297,7 +297,7 @@ private:
     uint16_t filterCount;
     uint16_t projectCount;
     bool needRidsAtDelivery;
-    uint8_t ridMap;
+    uint16_t ridMap;
 
 //	TableBand templateTB;
     uint32_t tableOID;
